refactor(canvas): remove RichChatInput component
- Deleted the RichChatInput component to streamline the codebase and improve maintainability.
- Ensured adherence to Tailwind CSS styling conventions and safe property access practices throughout the code.
